requirements.txt
numpy
opencv-python
Pillow
PyOpenGL
pyqt5
